move main news component to its own folder
