+++ src/components/News/mainNews/newsItem.scss
